Move PropertyValue into separate files :truck:
