+++ Src/Notion.Client/Models/PropertyValue/CheckboxPropertyValue.cs
@@ -1,0 +1,10 @@
+﻿namespace Notion.Client
+{
+    public class CheckboxPropertyValue : PropertyValue
+    {
+        public override PropertyValueType Type => PropertyValueType.Checkbox;
+
+        public bool Checkbox { get; set; }
+    }
+
+}
