+++ Src/Notion.Client/Models/PropertyValue/EmailPropertyValue.cs
@@ -1,0 +1,10 @@
+﻿namespace Notion.Client
+{
+    public class EmailPropertyValue : PropertyValue
+    {
+        public override PropertyValueType Type => PropertyValueType.Email;
+
+        public string Email { get; set; }
+    }
+
+}
