+++ Src/Notion.Client/Models/PropertyValue/LastEditedByPropertyValue.cs
@@ -1,0 +1,13 @@
+﻿using Newtonsoft.Json;
+
+namespace Notion.Client
+{
+    public class LastEditedByPropertyValue : PropertyValue
+    {
+        public override PropertyValueType Type => PropertyValueType.LastEditedBy;
+
+        [JsonProperty("last_edited_by")]
+        public User LastEditedBy { get; set; }
+    }
+
+}
