+++ Src/Notion.Client/Models/PropertyValue/NumberPropertyValue.cs
@@ -1,0 +1,9 @@
+﻿namespace Notion.Client
+{
+    public class NumberPropertyValue : PropertyValue
+    {
+        public override PropertyValueType Type => PropertyValueType.Number;
+        public double Number { get; set; }
+    }
+
+}
