+++ Src/Notion.Client/Models/PropertyValue/PeoplePropertyValue.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace Notion.Client
+{
+    public class PeoplePropertyValue : PropertyValue
+    {
+        public override PropertyValueType Type => PropertyValueType.People;
+        public List<User> People { get; set; }
+    }
+
+}
