+++ Src/Notion.Client/Models/PropertyValue/RichTextPropertyValue.cs
@@ -1,0 +1,14 @@
+﻿using System.Collections.Generic;
+using Newtonsoft.Json;
+
+namespace Notion.Client
+{
+    public class RichTextPropertyValue : PropertyValue
+    {
+        public override PropertyValueType Type => PropertyValueType.RichText;
+
+        [JsonProperty("rich_text")]
+        public List<RichTextBase> RichText { get; set; }
+    }
+
+}
