+++ Src/Notion.Client/Models/PropertyValue/SelectPropertyValue.cs
@@ -1,0 +1,9 @@
+﻿namespace Notion.Client
+{
+    public class SelectPropertyValue : PropertyValue
+    {
+        public override PropertyValueType Type => PropertyValueType.Select;
+        public SelectOption Select { get; set; }
+    }
+
+}
