+++ Src/Notion.Client/Models/PropertyValue/TitlePropertyValue.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace Notion.Client
+{
+    public class TitlePropertyValue : PropertyValue
+    {
+        public override PropertyValueType Type => PropertyValueType.Title;
+        public List<RichTextBase> Title { get; set; }
+    }
+
+}
