+++ Src/Notion.Client/Models/PropertyValue/UrlPropertyValue.cs
@@ -1,0 +1,10 @@
+﻿namespace Notion.Client
+{
+    public class UrlPropertyValue : PropertyValue
+    {
+        public override PropertyValueType Type => PropertyValueType.Url;
+
+        public string Url { get; set; }
+    }
+
+}
